chore(RHINENG-20838): Clean-up unused Edge components
